requirements.txt
pyjq==2.3.1
boto3==1.16.13
pytz==2020.1
PyYAML==5.3.1
config==0.4.2
psycopg2-binary==2.8.5
dataclasses-json==0.5.2
sqlalchemy==1.3.19
backoff==1.10.0
netaddr==0.7.19
requests==2.22.0
arnparse==0.0.2
pyjwt==1.7.1
packaging==20.4